requirements.txt
# Core dependencies
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
easyocr>=1.7.0
pandas>=2.0.0
pillow-heif>=0.13.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
